Unique filenames for environment files more explicit
I think it would be more helpful to make it clear that `environment.yml` doesn't have to always be used and the
environment files can be called anything. This avoids the possibility of easily over-writing existing files.

I've seen on other projects that people just go with what is in the documentation and create `sample-config.yaml`, which
isn't that helpful as they should be unique to a project so people don't have 384 files all called the same thing.